:lock: move test creds constants in a separate file
