+++ packages/hub/src/consts.ts
@@ -1,4 +1,1 @@
 export const HUB_URL = "https://huggingface.co";
-export const TEST_HUB_URL = "https://hub-ci.huggingface.co";
-export const TEST_USER = "hub.js";
-export const TEST_ACCESS_TOKEN = "hf_hub.js";
