+++ packages/hub/src/lib/upload-file.spec.ts
@@ -1,6 +1,6 @@
 import { assert, it, describe } from "vitest";
 
-import { TEST_ACCESS_TOKEN, TEST_HUB_URL, TEST_USER } from "../consts";
+import { TEST_ACCESS_TOKEN, TEST_HUB_URL, TEST_USER } from "../test/consts";
 import type { RepoId } from "../types/public";
 import { insecureRandomString } from "../utils/insecureRandomString";
 import { createRepo } from "./create-repo";
