+++ packages/hub/src/test/consts.ts
@@ -1,0 +1,3 @@
+export const TEST_HUB_URL = "https://hub-ci.huggingface.co";
+export const TEST_USER = "hub.js";
+export const TEST_ACCESS_TOKEN = "hf_hub.js";
